packed: Rework public API
 * Change image trait nameing
 * Adjust packed struct names accordingly
 * Export generic types in the prelude mod

We chose a nameing similar to the popular image trait so users do not need to
take a deep dive into the docs to get a hint as to what a struct/trait does.
On top of that, "ViewMut" should be a better term in the Rust world compared
to "FlatBuffer" although those convey the same meaning.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>