Get rid of image types
They are not needed to accomplish the main task of this crate:
converting between various pixelformats using arbitrary buffers.

We now place the burden of choosing a suitable storage type onto
the caller. The caller can then iterate through the collection and
map each pixel individually (or multiple, in the case of macropixels).

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>