yuv420: Fix plane packing
There is no way to guess width and height from the length of the buffer,
except for very specific cases (e.g. a 6x4 image, where the length of the
U and V planes matches the width by accidence).

There was another bug where the i (row) and j (column) iterators were
zipped, which lead to them increasing in lockstep. Instead, we want a
nested loop.

Signed-off-by: Christopher N. Hesse <raymanfx@gmail.com>